+++ src/storage/test/ScanVertexTest.cpp
@@ -17,7 +17,7 @@ cpp2::ScanVertexRequest buildRequest(
         PartitionID partId,
         const std::string& cursor,
         const std::pair<TagID, std::vector<std::string>>& tag,
-        int32_t rowLimit = 100,
+        int64_t rowLimit = 100,
         int64_t startTime = 0,
         int64_t endTime = std::numeric_limits<int64_t>::max(),
         bool onlyLatestVer = false) {
